ajustes importantes
