+++ app/build.gradle.kts
@@ -81,7 +81,7 @@ dependencies {
     implementation("androidx.appcompat:appcompat:1.7.0")
     implementation ("com.google.maps.android:maps-compose:2.2.0")
     implementation("androidx.room:room-common:2.6.1")
-
+    implementation("com.google.accompanist:accompanist-flowlayout:0.28.0") // O la última versión disponible
     implementation ("androidx.room:room-runtime:2.4.3")
     implementation("androidx.compose.runtime:runtime-livedata:1.7.4")
     kapt ("androidx.room:room-compiler:2.5.1") // Procesador de anotaciones para Room
