--- app/src/main/java/com/uniandes/ecobites/ui/screens/home/CategoriesRow.kt
@@ -1,62 +1,0 @@
-import androidx.compose.foundation.layout.*
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.rounded.Check
-import androidx.compose.material.icons.rounded.Edit
-import androidx.compose.material.icons.rounded.Home
-import androidx.compose.material.icons.rounded.ShoppingCart
-import androidx.compose.material.icons.rounded.Star
-import androidx.compose.material3.*
-import androidx.compose.runtime.*
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.vector.ImageVector
-import androidx.compose.ui.text.style.TextOverflow
-import androidx.compose.ui.unit.dp
-
-// Categories for the Tabs with Icons from Material Icons
-data class Category(val name: String, val icon: ImageVector)
-
-val categories = listOf(
-    Category("Restaurant", Icons.Rounded.Home),
-    Category("Ingredients", Icons.Rounded.Edit),
-    Category("Store", Icons.Rounded.ShoppingCart),
-    Category("Diet", Icons.Rounded.Check)
-)
-
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-fun CategoriesRow() {
-    var selectedTabIndex by remember { mutableStateOf(0) }
-
-    ScrollableTabRow(
-        selectedTabIndex = selectedTabIndex,
-        modifier = Modifier.fillMaxWidth(),
-        containerColor = MaterialTheme.colorScheme.surface,
-        contentColor = MaterialTheme.colorScheme.onSurface,
-        edgePadding = 16.dp // Optional padding at the edges
-    ) {
-        categories.forEachIndexed { index, category ->
-            Tab(
-                selected = selectedTabIndex == index,
-                onClick = { selectedTabIndex = index },
-                text = {
-                    // Row to place the icon and text horizontally
-                    Row(
-                        verticalAlignment = androidx.compose.ui.Alignment.CenterVertically
-                    ) {
-                        Icon(
-                            imageVector = category.icon,
-                            contentDescription = "${category.name} Icon",
-                            modifier = Modifier.size(20.dp)
-                        )
-                        Spacer(modifier = Modifier.width(8.dp)) // Space between icon and text
-                        Text(
-                            text = category.name,
-                            maxLines = 1,
-                            overflow = TextOverflow.Ellipsis
-                        )
-                    }
-                }
-            )
-        }
-    }
-}
