+++ app/src/main/java/com/uniandes/ecobites/ui/screens/restaurants/ExitoScreen.kt
@@ -1,0 +1,2 @@
+package com.uniandes.ecobites.ui.screens.restaurants
+
